check if webserv is running: not sure if it works on mac

+++ tester-webserv/run.sh
@@ -10,7 +10,15 @@ if [ "$?" != "0" ]; then
 	echo "Makefile Error"
 	exit 1
 fi
+
 $WEBSERV_DIR/$EXEC_NAME $CONFIG_FILE > /dev/null 2>&1 &
+
+process=$(ps -C ${EXEC_NAME} | grep a.out)
+if [ -z "$process" ]; then
+	echo "ERROR: not running: $EXEC_NAME"
+	exit 1
+fi
+
 make > /dev/null && python3 src/main.py $IP_PORT $@
 
 pkill $EXEC_NAME
